src/soc_llm_policy/pygraph_report.py
from __future__ import annotations

import argparse
import json
from dataclasses import dataclass
from pathlib import Path
from typing import Any


@dataclass(frozen=True)
class ModelViolationPoint:
    model_label: str
    violation_rate: float
    ci_low: float
    ci_high: float


@dataclass(frozen=True)
class RuleViolationPoint:
    model_label: str
    rule_id: str
    violation_count: int


def _load_bundle(path: Path) -> dict[str, Any]:
    data = json.loads(path.read_text(encoding="utf-8"))
    if not isinstance(data, dict):
        raise ValueError(f"Invalid bundle: {path}")
    return data


def _extract_model_violation_points(
    bundle: dict[str, Any],
) -> list[ModelViolationPoint]:
    rows = bundle.get("by_model", [])
    out: list[ModelViolationPoint] = []
    for row in rows:
        if not isinstance(row, dict):
            continue
        out.append(
            ModelViolationPoint(
                model_label=str(row.get("model_label", "unknown")),
                violation_rate=float(row.get("run_violation_rate", 0.0)),
                ci_low=float(row.get("run_violation_rate_ci_low", 0.0)),
                ci_high=float(row.get("run_violation_rate_ci_high", 0.0)),
            )
        )
    return out


def _extract_rule_violation_points(bundle: dict[str, Any]) -> list[RuleViolationPoint]:
    rows = bundle.get("by_rule", [])
    out: list[RuleViolationPoint] = []
    for row in rows:
        if not isinstance(row, dict):
            continue
        out.append(
            RuleViolationPoint(
                model_label=str(row.get("model_label", "unknown")),
                rule_id=str(row.get("rule_id", "unknown")),
                violation_count=int(row.get("violation_count", 0)),
            )
        )
    return out


def _extract_attack_tactic_counts(bundle: dict[str, Any]) -> dict[str, int]:
    summary = bundle.get("summary", {})
    if not isinstance(summary, dict):
        return {}
    raw = summary.get("attack_tactic_counts", {})
    if not isinstance(raw, dict):
        return {}
    out: dict[str, int] = {}
    for key, value in raw.items():
        out[str(key)] = int(value)
    return out


def _extract_token_cost(bundle: dict[str, Any]) -> tuple[int, float]:
    summary = bundle.get("summary", {})
    if not isinstance(summary, dict):
        return 0, 0.0
    tokens = int(summary.get("llm_total_tokens_total", 0))
    cost = float(summary.get("llm_cost_estimated_usd_total", 0.0))
    return tokens, cost


def _lazy_import_pyplot() -> Any:
    try:
        import matplotlib.pyplot as plt  # noqa: PLC0415
    except ImportError as exc:  # pragma: no cover - environment without matplotlib
        raise RuntimeError(
            "matplotlib is not installed. Run: .venv/bin/pip install matplotlib"
        ) from exc
    return plt


def _plot_violation_rate(points: list[ModelViolationPoint], outpath: Path) -> None:
    if not points:
        return
    plt = _lazy_import_pyplot()

    labels = [point.model_label for point in points]
    rates = [point.violation_rate for point in points]
    err_low = [max(point.violation_rate - point.ci_low, 0.0) for point in points]
    err_high = [max(point.ci_high - point.violation_rate, 0.0) for point in points]

    fig, ax = plt.subplots(figsize=(9, 5))
    ax.bar(labels, rates, color="#3b82f6", alpha=0.85)
    ax.errorbar(
        labels,
        rates,
        yerr=[err_low, err_high],
        fmt="none",
        ecolor="#111827",
        capsize=5,
    )
    ax.set_ylim(0, 1)
    ax.set_ylabel("Run Violation Rate")
    ax.set_title("Violation Rate by Model (95% Wilson CI)")
    ax.grid(axis="y", alpha=0.3)
    fig.tight_layout()
    fig.savefig(outpath, dpi=150)
    plt.close(fig)


def _plot_rule_breakdown(points: list[RuleViolationPoint], outpath: Path) -> None:
    if not points:
        return
    plt = _lazy_import_pyplot()

    rules = sorted({point.rule_id for point in points})
    models = sorted({point.model_label for point in points})
    series: dict[str, list[int]] = {
        model: [0 for _ in rules] for model in models
    }
    idx_by_rule = {rule: i for i, rule in enumerate(rules)}
    for point in points:
        series[point.model_label][idx_by_rule[point.rule_id]] += point.violation_count

    width = 0.8 / max(len(models), 1)
    x = list(range(len(rules)))

    fig, ax = plt.subplots(figsize=(10, 5))
    for pos, model in enumerate(models):
        offset = (pos - (len(models) - 1) / 2) * width
        shifted = [value + offset for value in x]
        ax.bar(shifted, series[model], width=width, label=model)

    ax.set_xticks(x)
    ax.set_xticklabels(rules)
    ax.set_ylabel("Violation Count")
    ax.set_title("Violations by Rule and Model")
    ax.legend()
    ax.grid(axis="y", alpha=0.3)
    fig.tight_layout()
    fig.savefig(outpath, dpi=150)
    plt.close(fig)


def _plot_attack_tactics(tactic_counts: dict[str, int], outpath: Path) -> None:
    if not tactic_counts:
        return
    plt = _lazy_import_pyplot()

    labels = list(tactic_counts.keys())
    values = list(tactic_counts.values())

    fig, ax = plt.subplots(figsize=(10, 5))
    ax.bar(labels, values, color="#10b981", alpha=0.85)
    ax.set_ylabel("Match Count")
    ax.set_title("ATT&CK Tactic Frequency")
    ax.grid(axis="y", alpha=0.3)
    fig.tight_layout()
    fig.savefig(outpath, dpi=150)
    plt.close(fig)


def _plot_budget(tokens: int, cost_usd: float, outpath: Path) -> None:
    plt = _lazy_import_pyplot()
    fig, ax = plt.subplots(figsize=(6, 4))
    labels = ["Total Tokens", "Estimated Cost (USD)"]
    values = [float(tokens), float(cost_usd)]
    ax.bar(labels, values, color=["#6366f1", "#f59e0b"])
    ax.set_title("Experiment Usage Budget")
    ax.grid(axis="y", alpha=0.3)
    fig.tight_layout()
    fig.savefig(outpath, dpi=150)
    plt.close(fig)


def build_parser() -> argparse.ArgumentParser:
    parser = argparse.ArgumentParser(prog="soc_llm_policy.pygraph_report")
    parser.add_argument(
        "--analysis-bundle-json",
        default="results/analysis/analysis_bundle.json",
        help="Typed bundle generated by analyze.py.",
    )
    parser.add_argument(
        "--outdir",
        default="results/analysis/figures",
        help="Output directory for PNG figures.",
    )
    return parser


def main(argv: list[str] | None = None) -> None:
    args = build_parser().parse_args(argv)
    bundle_path = Path(args.analysis_bundle_json).resolve()
    outdir = Path(args.outdir).resolve()
    outdir.mkdir(parents=True, exist_ok=True)

    bundle = _load_bundle(bundle_path)
    model_points = _extract_model_violation_points(bundle)
    rule_points = _extract_rule_violation_points(bundle)
    tactic_counts = _extract_attack_tactic_counts(bundle)
    tokens, cost_usd = _extract_token_cost(bundle)

    paths = {
        "violation_rate": outdir / "violation_rate_by_model.png",
        "rule_breakdown": outdir / "violations_by_rule_model.png",
        "attack_tactics": outdir / "attack_tactic_frequency.png",
        "budget": outdir / "usage_budget.png",
    }
    _plot_violation_rate(model_points, paths["violation_rate"])
    _plot_rule_breakdown(rule_points, paths["rule_breakdown"])
    _plot_attack_tactics(tactic_counts, paths["attack_tactics"])
    _plot_budget(tokens, cost_usd, paths["budget"])

    print(f"Figures saved at: {outdir}")
    for name, path in paths.items():
        if path.exists():
            print(f" - {name}: {path}")


if __name__ == "__main__":
    main()
