requirements.txt
# Runtime dependencies
pydantic>=2.7
pyyaml>=6.0
python-dotenv>=1.0
